requirements.txt
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.4
opencv-python==4.7.0.68
packaging==21.3
pandas==1.4.3
pytorch-model-summary==0.1.2
scikit-learn==1.3.2
scipy==1.8.1
seaborn==0.11.2
